requirements.txt
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.7.0 ; python_version >= "3.10" and python_version < "4"
cachetools==5.3.3 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4"
cloudevents==1.10.1 ; python_version >= "3.10" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
deprecation==2.1.0 ; python_version >= "3.10" and python_version < "4"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
flask==3.0.2 ; python_version >= "3.10" and python_version < "4"
functions-framework==3.5.0 ; python_version >= "3.10" and python_version < "4"
google-ai-generativelanguage==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.18.0 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.18.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.29.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-firestore==2.15.0 ; python_version >= "3.10" and python_version < "4.0"
google-generativeai==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.63.0 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.62.1 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.62.1 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.3 ; python_version >= "3.10" and python_version < "4"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4"
packaging==24.0 ; python_version >= "3.10" and python_version < "4"
proto-plus==1.23.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic-core==2.16.3 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.6.4 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.66.2 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
watchdog==4.0.0 ; python_version >= "3.10" and python_version < "4"
werkzeug==3.0.1 ; python_version >= "3.10" and python_version < "4"
